rename: JSX를 반환하는 파일들 .jsx로 변경 (#4)
