+++ src/app/(auth)/sign-up/page.jsx
@@ -1,16 +1,14 @@
-import styled from 'styled-components';
-
-const Login = () => {
+const SignUp = () => {
     return (
         <div>
-            <h1>Login Page</h1>
+            <h1>SignUp Page</h1>
             <form>
                 <input type="email" placeholder="Email" />
                 <input type="password" placeholder="Password" />
-                <button type="submit">Login</button>
+                <button type="submit">SignUp</button>
             </form>
         </div>
     );
 };
 
-export default Login;
+export default SignUp;
