+++ src/app/(main)/help-mungz/page.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const HelpMungz = () => {
     return (
         <div>
