+++ src/app/(main)/home/mbti/page.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const MbtiMatch = () => {
     return (
         <div>
