+++ src/app/(main)/mypage/edit/page.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const EditInfo = () => {
     return (
         <div>
