+++ src/app/(main)/mypage/page.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const MyPage = () => {
     return (
         <div>
