+++ src/app/match-info/page.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const MatchInfo = () => {
     return (
         <div>
