+++ src/app/onboarding/page.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const Onboarding = () => {
     return (
         <div>
