+++ src/app/providers.jsx
@@ -1,9 +1,10 @@
 'use client';
 
-import { useState } from 'react';
 import { useServerInsertedHTML } from 'next/navigation';
+import { useState } from 'react';
 import { ServerStyleSheet, StyleSheetManager } from 'styled-components';
-import GlobalStyle from '@/styles/globalStyles';
+
+import GlobalStyle from '../styles/globalStyles';
 
 export function Providers({ children }) {
     // ServerStyleSheet을 한 번만 생성 (성능 최적화)
